feat: 展示模式添加全屏水印
