+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "dev": "vite",
     "dev-host": "vite --host",
-    "build": "vite build && vue-tsc --noEmit",
+    "build": "vue-tsc --noEmit && vite build",
     "preview": "vite preview",
     "build-only": "vite build",
     "lint": "vue-tsc --noEmit && eslint . --ext .vue,.js,.ts,.jsx,.tsx --fix"
